+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 Changelog for the dpdumper utility
 
-## [0.0.3] - ????-??-??
+## [0.0.3] - 2024-07-29
+### Changed
+- Depends on dupicolib 0.2.0
+
 ### Added
 - Print the adapter notes when present, and allow them to be skipped
 - Add timed notice before starting the self-test
+- Read firmware version from board
+- Support different command classes depending on model and fw version
 
 ## [0.0.2] - 2024-07-26
 - Initial release